fix: 显式cast

+++ src/engine/core/transform.cpp
@@ -18,8 +18,8 @@ constexpr Matrix3f make_translate_matrix(float x, float y) {
 }
 
 constexpr Matrix3f make_rotate_matrix(float radian) {
-    const auto c = cos(radian);
-    const auto s = sin(radian);
+    const auto c = static_cast<float>(cos(radian));
+    const auto s = static_cast<float>(sin(radian));
     // c -s  0
     // s  c  0
     // 0  0  1
